Add SceneDelegate and update Info.plist for scene management
